requirements.txt
absl-py==0.11.0
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-oauthlib==0.4.3
six>=1.13.0